# Speed up method patching with identity Hashes