### Motivation

The `@sigs_that_raised` and `@was_ever_final_names` hashes are keyed by types for which `#==` is the same as `Object#equal?` (`Symbol`s and `Module`s). We can speed these up by using [`Hash#compare_by_identity`](https://ruby-doc.org/3.2.2/Hash.html#method-i-compare_by_identity).

Identity hashes are about 1.3x faster at the moment (in Ruby 3.2.1):

```
Warming up --------------------------------------
 regular Hash            1.520M i/100ms
 compare_by_identity     1.948M i/100ms
Calculating -------------------------------------
 regular Hash            15.122M (± 1.4%) i/s -     76.002M in   5.026956s
 compare_by_identity     19.482M (± 1.2%) i/s -     97.422M in   5.001500s

Comparison:
 compare_by_identity: 19481691.7 i/s
 regular Hash       : 15121932.4 i/s - 1.29x  slower
```

<details><summary><code>benchmark.rb</code></summary>

```ruby
#!/usr/bin/ruby

require 'benchmark/ips'

k = Module.new
hash = { k => true }
hash_by_id = { k => true }.compare_by_identity

Benchmark.ips do |x|
  x.report("regular Hash       ") { hash[k.object_id] }
  x.report("compare_by_identity") { hash_by_id[k]     }
  x.compare!
end
```

</details> 

### Test plan
<!-- If you did not write tests for this change, replace the message below explaining why not. Why we should be confident this change is correct? If you changed the website, please include a screenshot of the proposed changes. -->

See included automated tests.