@froydnj 2023-12-12
This is very cool!  Verified the same speedup on 2.7 and 3.1.