@froydnj 2023-12-13
We have a policy of testing changes to Sorbet against Stripe's codebase before
merging them. I've kicked off a test run for the current PR.  When the build
finishes, I'll share with you whether or how it failed. Thanks!

Stripe employees can see the build results here:

→ https://go/builds/bui_PBFpggwCjFKUkA
→ https://go/builds/bui_PBFpPBCPtYaKuz
→ https://go/builds/bui_PBFpMuyDRxSkJ8