@froydnj 2023-12-14
> Hey @froydnj is this good to merge?

This looks OK to me, but testing it is running into some bitrot with our sorbet-runtime test scripts (since it's been a while since we've had to test sorbet-runtime) and I'm not entirely sure how long the issues will take to fix.